% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regRCPqn_regRCPqnREF.R
\name{regRCPqn}
\alias{regRCPqn}
\title{The regRCPqn Function}
\arguments{
\item{M_data:}{data.frame containing M-values. The first column of M_data has to be named ID_REF and to contain the CpG codes. Each other column corresponds to a sample.}

\item{ref_path:}{path to folder in which files will be saved}

\item{data_name:}{prefix that will be used to create the file name of the reference distribuion if save_ref is TRUE.}

\item{save_ref}{(TRUE): whether to save the reference distribution of the normalized data set.}
}
\description{
The regRCPqn Function
}
\examples{
regRCPqn(M_data, ref_path, data_name,save_ref=TRUE)
}
