% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regRCPqn_regRCPqnREF.R
\name{regRCPqnREF}
\alias{regRCPqnREF}
\title{The regRCPqnREF Function}
\usage{
regRCPqnREF(M_data, ref_path, data_name)
}
\arguments{
\item{M_data:}{data.frame containing M-values. The first column of M_data has to be named ID_REF and to contain the CpG codes. Each other column corresponds to a sample.}

\item{ref_path:}{path to folder in which the reference distribution of the genomic region types computed with regRCPqn have been saved.}

\item{data_name:}{data_name: prefix used in regRCPqn to save the reference distribution files.}
}
\description{
The regRCPqnREF Function
}
\examples{
# Load data
data(BloodParkinson2)
# Set directory to save reference distribution
ref_path <- "./example/"
# Set dataset label
data_name <- "Example"
# Run regRCPqnREF
M_data_norm_ref <- regRCPqnREF(M_data=BloodParkinson2, ref_path=ref_path, data_name=data_name)
}
